Run vm remove_from_disk via MiqQueue
https://bugzilla.redhat.com/show_bug.cgi?id=1183167

Previously the removal of a vm needed the Automate Role and VIM broker
on the same server. This PR allows the vm remove_from_disk to be
routed to the appropriate worker which has the ems_operations role.
