+++ vmdb/lib/miq_automation_engine/service_models/miq_ae_service_vm_vmware.rb
@@ -1,5 +1,8 @@
 module MiqAeMethodService
   class MiqAeServiceVmVmware < MiqAeServiceVmInfra
+    require_relative "mixins/miq_ae_service_ems_operations_mixin"
+    include MiqAeServiceEmsOperationsMixin
+
     def set_number_of_cpus(count, options = {})
       sync_or_async_ems_operation(options[:sync], "set_number_of_cpus", [count])
     end
@@ -10,32 +13,6 @@ module MiqAeMethodService
 
     def add_disk(disk_name, disk_size_mb, options = {})
       sync_or_async_ems_operation(options[:sync], "add_disk", [disk_name, disk_size_mb])
-    end
-
-    private
-
-    def sync_or_async_ems_operation(sync, method_name, args)
-      ar_method do
-        queue_options = {
-          :class_name  => @object.class.name,
-          :instance_id => @object.id,
-          :method_name => method_name,
-          :args        => args,
-          :zone        => @object.my_zone,
-          :role        => "ems_operations",
-          :task_id     => nil               # Clear task_id to allow running synchronously under current worker process
-        }
-
-        if sync
-          task_options = {:name => method_name.titleize, :userid => "system"}
-          task_id = MiqTask.generic_action_with_callback(task_options, queue_options)
-          MiqTask.wait_for_taskid(task_id)
-        else
-          MiqQueue.put(queue_options)
-        end
-      end
-
-      true
     end
   end
 end
